auto_wchat_basic_project_20260101/requirements.txt
# Web自动化
selenium==4.15.2
webdriver-manager==4.0.1

# 环境配置
python-dotenv==1.0.0

# AI服务
openai==1.3.0
dashscope>=1.14.0

# HTTP客户端
httpx==0.27.2
urllib3>=2.0.0

# 翻译服务
deep-translator==1.11.4

# 任务调度
schedule==1.2.0

# 二维码生成
qrcode==7.4.2
Pillow==10.1.0

# Web框架
flask==3.0.0
flask-cors==4.0.0

# 文件处理
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3

# Optional: Support for .doc format (old Word documents)
# textract>=1.6.5  # Uncomment to support .doc format
# pywin32>=306  # Windows only, uncomment to support .doc format
